harden(rbaw-v1-t004): bind evaluation integrity
