--- .github/workflows/rbaw-hardening-pr-bootstrap.yml
@@ -1,80 +1,0 @@
-name: RBAW hardening PR bootstrap
-
-on:
-  pull_request:
-    branches:
-      - main
-
-concurrency:
-  group: rbaw-hardening-${{ github.event.pull_request.number }}
-  cancel-in-progress: true
-
-permissions:
-  contents: write
-
-jobs:
-  apply-and-verify:
-    if: github.head_ref == 'fix/rbaw-v1-t004-integrity-hardening-v2'
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    steps:
-      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-        with:
-          fetch-depth: 0
-          persist-credentials: true
-
-      - name: Set up Python
-        uses: actions/setup-python@5fda3b95a4ea91299a34e894583c3862153e4b97
-        with:
-          python-version: "3.12"
-
-      - name: Install dependencies and Bubblewrap
-        run: |
-          python -m pip install --disable-pip-version-check --require-hashes -r requirements/repoground-dev.lock.txt
-          sudo apt-get update
-          sudo apt-get install --no-install-recommends --yes bubblewrap
-          if sysctl kernel.unprivileged_userns_clone >/dev/null 2>&1; then
-            sudo sysctl -w kernel.unprivileged_userns_clone=1
-          fi
-          if sysctl kernel.apparmor_restrict_unprivileged_userns >/dev/null 2>&1; then
-            sudo sysctl -w kernel.apparmor_restrict_unprivileged_userns=0
-          fi
-
-      - name: Apply integrity hardening
-        run: |
-          python - <<'PY'
-          import base64
-          import gzip
-          import re
-          from pathlib import Path
-
-          workflow = Path('.github/workflows/rbaw-hardening-bootstrap.yml').read_text(encoding='utf-8')
-          match = re.search(r'TRANSFORM_GZIP_B64: "([A-Za-z0-9+/=]+)"', workflow)
-          if match is None:
-              raise SystemExit('embedded transformer payload is missing')
-          payload = gzip.decompress(base64.b64decode(match.group(1)))
-          Path('.github/apply_rbaw_hardening.py').write_bytes(payload)
-          PY
-          python .github/apply_rbaw_hardening.py
-          rm .github/apply_rbaw_hardening.py
-
-      - name: Format and verify focused suite
-        run: |
-          python -m ruff format tools/patch_evaluation_sidecar.py tests/test_patch_evaluation_sidecar.py
-          python -m ruff check tools/patch_evaluation_sidecar.py tests/test_patch_evaluation_sidecar.py
-          python -m pytest -q tests/test_patch_evaluation_sidecar.py merger/repoground/tests/test_patch_evaluation.py
-          python -m py_compile tools/patch_evaluation_sidecar.py tests/test_patch_evaluation_sidecar.py
-          git diff --check
-
-      - name: Verify full suite
-        run: pytest -q -m "not browser and not doc_freshness_live"
-
-      - name: Commit verified hardening
-        run: |
-          rm .github/workflows/rbaw-hardening-bootstrap.yml
-          rm .github/workflows/rbaw-hardening-pr-bootstrap.yml
-          git config user.name "grabowski-bot"
-          git config user.email "alexdermohr+bot@users.noreply.github.com"
-          git add -A
-          git commit -m "harden(rbaw-v1-t004): bind evaluation integrity"
-          git push origin HEAD:fix/rbaw-v1-t004-integrity-hardening-v2
